Add test for wizard resolver Co-authored-by: Joanna Fawl <joannafawl@users.noreply.github.com>
